requirements.txt
xlwt==1.3.0
requests==2.19.1
openpyxl==2.5.6
Pillow==5.3.0
tesserocr==2.3.1
